Fix deadlock when disconnecting connections made with function pointers
The regression was introduced in 5885b8f775998c30d53f40b7f368c5f6364e6df4

QMetaObjectPrivate::disconnectHelper may unlock the sender mutex. And
while relocking it, we need to make sure to lock the sender and receiver
mutex in the right order.  So don't lock the receiver mutex in advance, but
re-lock it for each connection.

Change-Id: I4f6d19791cdcce3693d7f45e7beb6b564fd69277
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Lars Knoll <lars.knoll@digia.com>