Add proper abstractions to the grid layout engine.
The abstractions are needed so that they can work with both
QGraphicsLayouts and QtQuick.Layouts.

Since the plan is to move the engine to QtGui, this means that the
engine cannot have any references to anything in the QtWidgets module.

As a consequence of that several things had to be done:
* The style info object had to be redone with an abstraction layer
  to get rid of style and widget dependency. (Abstract class is
  called QAbstractLayoutStyleInfo)
* QGridLayoutEngine must be subclassed due to some specializations for
  QGraphicsLayoutItem, manifested as QGraphicsGridLayoutEngine.
* QGridLayoutItem must be subclassed due to some specializations for
  QGraphicsLayoutItem, manifested as QGraphicsGridLayoutEngineItem.

Did also some minor cleanups, reordered arguments so that all styleInfo
arguments are last in all function calls

This also fixes QTBUG-35099 (bug was spotted during this refactoring)

Task-number: QTBUG-35099

Change-Id: If49d40f71870dc8d99d2e145be158e3080b595fa
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>