Fixes: 32 bit and alpha bmp format image reading
The bmp handler would ignore the high byte of a 32bit bmp. It
would also ignore any alpha channel.

The actual change is really simple; it just adds the alpha_* members
and the reading of the 4th byte in 32bit depth. However, detection
of alpha channel required switching the order of two independent
blocks of code, hence the size of this patch.

[ChangeLog][QtGui] Support reading bmp images with alpha channel

Task-number: QTBUG-35220
Change-Id: Ib133c3644c03c5cfc728f5fa2c837219804e69ae
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@digia.com>
Reviewed-by: Gunnar Sletta <gunnar.sletta@jollamobile.com>