QCups: avoid relocations
Replace a string (pointer) table with arrays of char arrays of maximally
occurring size.

In the pageLayoutData case, this is trivial, since all strings have the same size.

In the pagesPerSheetData case, I've used a trick to cram the (only) two-digit
number into a char[2] array, by following it with a null entry.

Effects on AMD64 Linux GCC 4.7 release stripped:
   text:   -352B
   data:   -160B
   relocs:  -14

Change-Id: I6c458ff7ada0f45dab976bbe42b24757fc321302
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>