QPlainTextEdit: set Qt::ImhMultiLine
Like QTextEdit, QPlainTextEdit should also set the
input method hint Qt::ImhMultiLine.
An important reason for this, is that the iOS plugin
will inspect the input method hints of the focus
object to decide which type of keyboard (and return
key) to use. Failing to set this hint will result
in the iOS plugin setting the return key to 'done'
instead of 'return' for this widget.

Change-Id: Idd5aaf21505add608f2621f94655984f42e92a18
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@digia.com>