Implement QMainWindow::setUnifiedTitleAndToolBarOnMac
Fix a feature regression from Qt 4.

Unlike the Qt 4 implementation this does not move
the tool bars out of the QMainWindow while pretending
they are still there.

Instead, use setContentBorderThickness from the Cocoa
platform plugin to draw a background gradient that
unifies the title and toolbar area.

QToolBar can then draw itself with a transparent
background and let the gradient shine throughout.

This is a style-only change, toolbar behavior is
kept as-is.

[ChangeLog][Platform Specific Changes][OS X] Implemented
QMainWindow::setUnifiedTitleAndToolBarOnMac.

Task-number: QTBUG-34411
Change-Id: Idcaab6399f249b11edb1147856d9aece9923ab36
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@digia.com>